Renaming helper files and final full split
